file rename + dir change
