+++ include/color_sequence.h
