+++ include/inital_calculations.h
